+++ src/iexec-orderbook.js
@@ -170,7 +170,6 @@ orderbookWorkerpool
 
       spinner.succeed(successMessage, {
         raw: {
-          openVolume: response.openVolume,
           workerpoolOrders: response.workerpoolOrders,
         },
       });